+++ magda-web-client/src/helpers/ChartDatasetEncoder.js
@@ -10,7 +10,7 @@ import concat from "lodash/concat";
 import takeRight from "lodash/takeRight";
 import sortBy from "lodash/sortBy";
 import * as d3 from "d3-collection";
-import chrono from "chrono-node";
+import * as chrono from "chrono-node";
 
 const AVAILABLE_CHART_TYPES = ["bar", "pie", "scatter", "line"];
 const STRIP_NUMBER_REGEX = /[^\-\d.+]/g;
@@ -69,24 +69,16 @@ const OVERRIDE_TO_NUMERIC_COLUMNS = [
 ];
 const EXCLUDE_COLUMNS = ["long", "lat", "lng"];
 
-const noDelimiterParser = new chrono.Parser();
-noDelimiterParser.pattern = function () {
-    return /(\d{2})(\d{2})(\d{4})/gi;
-};
-noDelimiterParser.extract = function (text, ref, match, opt) {
-    return new chrono.ParsedResult({
-        ref: ref,
-        text: match[1] + "/" + match[2] + "/" + match[3],
-        index: match.index,
-        start: {
-            day: match[1],
-            month: match[2],
-            year: match[3]
-        }
-    });
+const noDelimiterParser = {
+    pattern: () => /(\d{2})(\d{2})(\d{4})/gi,
+    extract: (context, match) => ({
+        day: match[1],
+        month: match[2],
+        year: match[3]
+    })
 };
 
-const customChrono = chrono.en_GB;
+const customChrono = chrono.en.GB.clone();
 customChrono.parsers.push(noDelimiterParser);
 
 const parseNumber = (str) => {
@@ -532,7 +524,7 @@ class ChartDatasetEncoder {
         };
 
         const unsortedData = inner().map((datum) => {
-            let rawValue = datum[this.xAxis.name];
+            const rawValue = datum[this.xAxis.name];
 
             if (getFieldDataType(this.xAxis) === "time") {
                 const parsedDate = customChrono.parseDate(rawValue);